utils/requirements.txt
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
numpy==2.2.6
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pycparser==3.0
PyMuPDF==1.27.1
pypdfium2==5.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.3
